+++ src/apis/cart.ts
@@ -1,10 +1,10 @@
-import { CartItemType } from "../types";
+import { CartItem } from "../types";
 import fetchWithAuth from "./utils/fetchWithAuth";
 
 /**
  * 사용자의 장바구니 목록 조회
  */
-export const getCartItems = async (): Promise<CartItemType[]> => {
+export const getCartItems = async (): Promise<CartItem[]> => {
   const response = await fetchWithAuth("/cart-items", {
     method: "GET",
   });
